Fix tag page + create concern to DRY tag methods
